remove single-use audit mixin
